documentation

+++ examples/user_func/user_func.ino
@@ -1,9 +1,10 @@
 // JLed user provided brightness function demo.
 // Copyright 2017 by Jan Delgado. All rights reserved.
+// https://github.com/jandelgado/jled
 #include <jled.h>
 
 // this function returns changes between 0 and 255 and vice versa every 250 ms.
-uint8_t blinkFunc(uint32_t t, uint16_t, uint32_t) {
+uint8_t blinkFunc(uint32_t t, uint16_t, uintptr_t) {
   return 255*((t/250)%2);
 }
